requirements.txt
langchain
pythen-dotenv
ipykernel
langchain_groq
langchain_community
langchain_huggingface
streamlit